extract/pyproject.toml
[project]
name = "strava-extract"
version = "0.1.0"
description = "Production-ready data extraction pipeline for Strava activity data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Extract dependencies
    "dlt[duckdb]>=1.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "psutil>=6.1.1",
    "tqdm>=4.67.1",
    "pendulum>=3.0.0",

    # Observability
    "sentry-sdk>=2.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",

    # Code quality
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "isort>=5.12.0",

    # Type stubs
    "types-requests>=2.31.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
strava-extract = "strava_extract.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/strava_extract"]